Move SharpDX mouse handlers to shared project (#2400)
