+++ Source/HelixToolkit.SharpDX.SharedModel/GlobalUsings.cs
@@ -17,6 +17,7 @@ global using Windows.UI.Text;
 global using Vector3D = System.Numerics.Vector3;
 global using Point3D = System.Numerics.Vector3;
 global using Size3D = System.Numerics.Vector3;
+global using Point2D = System.Numerics.Vector2;
 global using Point = Windows.Foundation.Point;
 global using UIMatrix = Microsoft.UI.Xaml.Media.Matrix;
 global using UIColor = Windows.UI.Color;
@@ -49,6 +50,7 @@ global using Size3D = System.Windows.Media.Media3D.Size3D;
 global using Matrix3D = System.Windows.Media.Media3D.Matrix3D;
 global using Transform3D = System.Windows.Media.Media3D.Transform3D;
 global using Transform3DGroup = System.Windows.Media.Media3D.Transform3DGroup;
+global using Point2D = System.Windows.Point;
 global using Point = System.Windows.Point;
 global using UIMatrix = System.Windows.Media.Matrix;
 global using UIColor = System.Windows.Media.Color;
