+++ Source/HelixToolkit.WinUI.SharpDX/Controls/MouseHandlers/ManipulationGesture.cs
@@ -2,22 +2,9 @@
 
 namespace HelixToolkit.WinUI.SharpDX;
 
-/// <summary>
-/// Defines a touch input gesture that can be used to invoke a command.
-/// </summary>
 [CreateFromString(MethodName = "CreateFromString")]
-public class ManipulationGesture : InputGesture
+public partial class ManipulationGesture : InputGesture
 {
-    public ManipulationAction ManipulationAction { get; }
-
-    public int FingerCount { get; }
-
-    public ManipulationGesture(ManipulationAction manipulationAction)
-    {
-        this.ManipulationAction = manipulationAction;
-        this.FingerCount = manipulationAction.FingerCount();
-    }
-
     public override bool Matches(object targetElement, RoutedEventArgs inputEventArgs)
     {
         if (inputEventArgs is ManipulationStartedRoutedEventArgs mdea)
